requirements.txt
langchain-openai==0.1.0
langchain-core==0.1.0
python-dotenv==1.0.0
openai==1.12.0

